Split out Register activity into two, with new activity for terms of service
The terms of service content comes from an HTML asset file, with the intention
that skinned apps can provide their own terms of service.

While I was refactoring, I also removed a bunch of dead code related to
uploading profile photos, which didn't work and had been disabled.

Connects to #271
