+++ OpenTreeMap/src/main/AndroidManifest.xml
@@ -100,8 +100,14 @@
             android:name="org.azavea.otm.ui.Register"
             android:label="@string/app_name"
             android:configChanges="orientation"
+            android:screenOrientation="portrait">
+        </activity>
+        <activity
+            android:name="org.azavea.otm.ui.TermsOfService"
+            android:label="@string/app_name"
+            android:configChanges="orientation"
             android:screenOrientation="portrait"
-            android:noHistory="true" >
+            android:noHistory="true">
         </activity>
         <activity
             android:name="org.azavea.otm.ui.TreeMove"
