+++ OpenTreeMap/src/main/assets/terms_of_service.html
@@ -1,0 +1,7 @@
+<html>
+<body>
+    <center>
+        Add terms of service here!
+    </center>
+</body>
+</html>
